# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Modeva CMS Backend API Documentation",
        "title": "Modeva CMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Retrieve a paginated list of all products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get all products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by subcategory",
                        "name": "sub_category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new product with Cloudinary URLs (optimized flow)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product details with Cloudinary URLs",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product_controller.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/products/{id}": {
            "get": {
                "description": "Retrieve a single product and its related details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update product details by ID with support for both text and image updates",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update an existing product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Product name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Product description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Product price",
                        "name": "price",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Sub category ID",
                        "name": "subCategoryId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Product status",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "New primary product image",
                        "name": "primaryImage",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Existing primary image URL (if not changing)",
                        "name": "primaryImageUrl",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "New other product images (multiple)",
                        "name": "otherImages",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "JSON array of existing other images to keep",
                        "name": "existingOtherImages",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Product variants JSON",
                        "name": "variants",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Product inventory JSON",
                        "name": "inventory",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "SEO details JSON",
                        "name": "seo",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Product composition JSON",
                        "name": "composition",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Product tags JSON",
                        "name": "tags",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a product by ID and its associated Cloudinary folder",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/store/products": {
            "get": {
                "description": "Get paginated products for storefront with optional search and filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Get storefront products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Category IDs (repeatable ?category=ID\u0026category=ID)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Subcategory IDs (repeatable ?subcategory=ID\u0026subcategory=ID)",
                        "name": "subcategory",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Sizes (repeatable ?size=XS\u0026size=S)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "in_stock",
                            "out_of_stock",
                            "inStock",
                            "outOfStock"
                        ],
                        "type": "string",
                        "description": "Availability filter",
                        "name": "availability",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum price",
                        "name": "minPrice",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum price",
                        "name": "maxPrice",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "price",
                            "name",
                            "newest",
                            "popular"
                        ],
                        "type": "string",
                        "default": "newest",
                        "description": "Sort by field",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/store/products/{id}": {
            "get": {
                "description": "Get detailed product information by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Get single product details for storefront",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/storefront/products": {
            "get": {
                "description": "Retrieve active storefront products with optional search, category, subcategory, size, colour, availability, price range, and sorting filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storefront - Products"
                ],
                "summary": "Get storefront products with filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (name or description)",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Parent category IDs (repeatable)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Subcategory IDs (repeatable)",
                        "name": "subcategory",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Sizes (repeatable)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Colours (repeatable)",
                        "name": "color",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Availability filter (in_stock | out_of_stock)",
                        "name": "availability",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum price",
                        "name": "minPrice",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum price",
                        "name": "maxPrice",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "newest",
                        "description": "Sort by field (newest, price, etc.)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (asc | desc)",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 12,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products with filters fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/models.Pagination"
                },
                "rate_limit": {
                    "$ref": "#/definitions/models.RateLimiter"
                },
                "requested_entity": {
                    "type": "string"
                }
            }
        },
...
        "models.MediaURL": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "order": {
                    "description": "only relevant for \"other\" images",
                    "type": "integer"
                },
                "url": {
                    "description": "primary or other must have this",
                    "type": "string"
                }
            }
        },
...
        "models.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 42
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "models.ProductMedia": {
            "type": "object",
            "required": [
                "primary"
            ],
            "properties": {
                "other": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaURL"
                    }
                },
                "primary": {
                    "$ref": "#/definitions/models.MediaURL"
                }
            }
        },
...
        "models.RateLimiter": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "remaining": {
                    "type": "integer"
                },
                "reset_at": {
                    "type": "string"
                },
                "reset_in_seconds": {
                    "type": "integer"
                }
            }
        },
...
        "product_controller.CreateProductRequest": {
            "type": "object",
            "required": [
                "description",
                "media",
                "name",
                "price",
                "subCategoryId"
            ],
            "properties": {
                "composition": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "inventory": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "media": {
                    "$ref": "#/definitions/models.ProductMedia"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "seo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "type": "string"
                },
                "subCategoryId": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}